requirements.txt
fastapi[all]
sqlalchemy[asyncio]
asyncpg
alembic
pydantic-settings
passlib[bcrypt]
python-jose[cryptography]
python-multipart
uvicorn
aiosqlite
alembic
argon2-cffi